fix babel polyfill for thirdparties #109
